Backfill announcement slugs in migration
Based on PR feedback, we backfill the announcement slugs in a migration
(instead of through a mix task) so that we can subsequently make the
slug column non-nullable. Ideally we would not use Elixir when updating
database values in a migration (since the modules could change and thus
ruin the migration in the future), but we need the
`Slugger.slugify_downcase/1` function to generate slugs.